requirements.txt
certifi==2020.12.5
Pillow==8.2.0
selenium==3.141.0
urllib3==1.26.4
